Re-register with parse server in test to Fix #91
The last 4 unit tests fail when run in series, but not if they were
run individually. The test for testCanGetCurrentUser registers the
service using the session token which loses the previous registration.
The other tests require that they are registered with the master_key,
so now the testCanGetCurrentUser test re-registers.